Fixed a couple of things and balanced out the gameplay at lower levels. THe player only has one life now, as the lives system will only be implemented in future updates.

+++ source/application.e
@@ -75,13 +75,15 @@ feature {NONE} -- Initialization
 			{SDL}.sdl_setrenderdrawcolor (l_window.renderer, 0, 0, 0, 255)
 			create l_event.make (l_window)
 			create l_title_screen.make (l_window, debug_mode)
-			l_event.destroy
-			l_window.destroy
 
-			if attached l_resources as la_resources then
-				l_resources.destroy
+			if not debug_mode then
+				if attached l_resources as la_resources then
+					l_resources.destroy
+				end
 			end
 
+			l_event.destroy
+			l_window.destroy
 			{SDL_MIXER}.mix_close_audio
 			{SDL_MIXER}.mix_quit
 			{SDL_TTF}.ttf_quit
